Refactor for packaging
