+++ pyproject.toml
@@ -1,0 +1,58 @@
+# Guide (user-friendly):
+# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
+
+# Specification (technical, formal):
+# https://packaging.python.org/en/latest/specifications/pyproject-toml/
+
+
+# Choosing a build backend:
+# https://packaging.python.org/en/latest/tutorials/packaging-projects/#choosing-a-build-backend
+[build-system]
+# A list of packages that are needed to build your package:
+requires = ["setuptools"] # REQUIRED if [build-system] table is used
+# The name of the Python object that frontends will use to perform the build:
+build-backend = "setuptools.build_meta" # If not defined, then legacy behavior can happen.
+
+[project]
+
+name = "tlscan"
+version = "3.0.1"
+description = "Python based TLS enumeration scanner"
+readme = "README.md"
+requires-python = ">=3.9"
+license = { file = "LICENSE.txt" }
+keywords = ["tls", "scanner", "security"]
+
+authors = [{ name = "M.H. Jansen of Lorkeers", email = "marco@threatlabs.eu" }]
+
+maintainers = [
+  { name = "ThreatLabs", email = "info@threatlabs.eu" },
+]
+
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+
+  "Intended Audience :: Information Technology",
+  # "Topic :: Security testing :: Build Tools",
+
+  "License :: OSI Approved :: MIT License",
+
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3 :: Only",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/ThreatLabsNL/TLScan"
+"Bug Reports" = "https://github.com/ThreatLabsNL/TLScan/issues"
+"Source" = "https://github.com/ThreatLabsNL/TLScan/"
+
+[tool.setuptools.packages.find]
+where = ["tlscan"]
+
+[project.scripts]
+tlscan = "tlscan:main"
