+++ tlscan/__init__.py
@@ -2,8 +2,8 @@ from sys import exit
 from argparse import ArgumentParser
 from datetime import datetime
 
-from scanner import TargetParser, Enumerator, start_tls
-from TLS.protocols import versions as p_versions
+from .scanner import TargetParser, Enumerator, start_tls
+from .TLS.protocols import versions as p_versions
 
 # ToDo
 # cipher preference
@@ -53,7 +53,7 @@ def main():
     parser = ArgumentParser(description='Scanner to enumerate encryption protocol support', prog='TLScan3')
     parser.add_argument('target', type=str, help="specify target as: host:port e.g. www.example.com:443 or "
                                                  "[::1]:443 for IPv6")
-    parser.add_argument('--version', action='version', version='%(prog)s 3.1')
+    parser.add_argument('--version', action='version', version='%(prog)s 3.1.1')
     p_group = parser.add_mutually_exclusive_group()
     for key, value in start_tls.items():
         p_group.add_argument("--{}".format(key), dest=key, action='store_true',
