+++ tlscan/__main__.py
@@ -1,0 +1,5 @@
+#! /usr/bin/env python
+
+import tlscan
+
+tlscan.main()
